Calculates the pooled C-index and Confidence intervals

pool_cindex Calculates the pooled C-index and Confidence intervals.

pool_cindex(data, conf.level = 0.95, dfcom = NULL)

Arguments

data

An object of class 'mistats' ('Multiply Imputed Statistical Analysis'.) or a m x 2 matrix with C-index values and standard errors in the columns. For the latter option dfcom has to be provided.

conf.level

conf.level Confidence level of the confidence intervals.

dfcom

Number of completed-data analysis degrees of freedom.

Value

The pooled c-index value and the confidence intervals.

Details

Rubin's Rules are used for pooling. The C-index values are log transformed before pooling and finally back transformed.

Vignettes

https://mwheymans.github.io/miceafter/articles/pooling_cindex.html

See also

Author

Martijn Heymans, 2021

Examples

 # Logistic Regression
 imp_dat <- df2milist(lbpmilr, impvar="Impnr")
 res_stats <- with(data=imp_dat,
  expr = cindex(glm(Chronic ~ Gender + Radiation,
  family=binomial)))
 res <- pool_cindex(res_stats)
 res
#>        C-index Critical value 95 CI low 95 CI high
#> [1,] 0.6638267       1.976656 0.5764274  0.7412861
#> attr(,"class")
#> [1] "mipool"

 # Cox regression
 library(survival)
 imp_dat <- df2milist(lbpmicox, impvar="Impnr")
  res_stats <- with(data=imp_dat,
  expr = cindex(coxph(Surv(Time, Status) ~ Pain + Radiation)))
 res <- pool_cindex(res_stats)
 res
#>        C-index Critical value 95 CI low 95 CI high
#> [1,] 0.5820398       1.976697 0.5330546  0.6294584
#> attr(,"class")
#> [1] "mipool"