Calculates the pooled Levene test.

pool_levenetest Calculates the pooled F-statistic of the Levenene test.

pool_levenetest(object, method = "D1")

Arguments

object

An object of class 'mistats' ('Multiply Imputed Statistical Analysis').

method

A character vector to choose the pooling method, 'D1' (default) or 'D2'.

Value

The (combined) F-statistic, p-value and degrees of freedom.

Vignettes

https://mwheymans.github.io/miceafter/articles/levene_test.html

References

Eekhout I, van de Wiel MA, Heymans MW. Methods for significance testing of categorical covariates in logistic regression models after multiple imputation: power and applicability analysis. BMC Med Res Methodol. 2017;17(1):129.

Enders CK (2010). Applied missing data analysis. New York: The Guilford Press.

Van Buuren S. (2018). Flexible Imputation of Missing Data. 2nd Edition. Chapman & Hall/CRC Interdisciplinary Statistics. Boca Raton.

See also

Author

Martijn Heymans, 2021

Examples

library(magrittr)
lbpmilr %>%
   df2milist(impvar="Impnr") %>%
      with(expr=levene_test(Pain ~ factor(Carrying))) %>%
        pool_levenetest(method="D1")
#>       F_value df1      df2    P(>F)       RIV
#> [1,] 1.586703   2 115.3418 0.209032 0.1809493
#> attr(,"class")
#> [1] "mipool"

# Same as
imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=levene_test(Pain ~ factor(Carrying)))
res <- pool_levenetest(ra, method="D1")