Calculates the pooled difference between proportions and confidence intervals according to Newcombe-Wilson (NW) across multiply imputed datasets.

pool_propdiff_nw Calculates the pooled difference between proportions and confidence intervals according to Newcombe-Wilson (NW) across multiply imputed datasets.

pool_propdiff_nw(object, conf.level = 0.95)

Arguments

object

An object of class 'mistats' ('Multiply Imputed Statistical Analysis'.).

conf.level

Confidence level of the confidence intervals. Mostly set at 0.95.

Value

The Proportion and the Confidence intervals according to Newcombe-Wilson.

Details

The pool_propdiff_nw function uses information from separate exposure groups. It is therefore important to first use the propdiff_wald function and to set strata = TRUE in that function.

References

Yulia Sidi & Ofer Harel (2021): Difference Between Binomial Proportions Using Newcombe’s Method With Multiple Imputation for Incomplete Data, The American Statistician, DOI:10.1080/00031305.2021.1898468

See also

Author

Martijn Heymans, 2021

Examples

imp_dat <- df2milist(lbpmilr, impvar="Impnr")
res <- with(imp_dat, expr=propdiff_wald(Chronic ~ Radiation, strata = TRUE))
res <- pool_propdiff_nw(res)
res
#>      Prop diff CI L NW CI U NW
#> [1,]    0.2786  0.1199   0.419
#> attr(,"class")
#> [1] "mipool"