Calculates the difference between proportions and standard error according to method Agresti-Caffo

propdiff_ac Calculates the difference between proportions and standard error according to method Agresti-Caffo.

Usage,

propdiff_ac(y, x, formula, data)

Arguments

y

0-1 binary response variable.

x

0-1 binary independent variable.

formula

A formula object to specify the model as normally used by glm.

data

An objects of class milist, created by df2milist, list2milist or mids2milist.

Value

The difference between proportions and the standard error according to Agresti-Caffo.

See also

Author

Martijn Heymans, 2021

Examples

imp_dat <- df2milist(lbpmilr, impvar="Impnr")
ra <- with(imp_dat, expr=propdiff_ac(Chronic ~ Radiation))